# swagger.yaml
openapi: 3.0.0
info:
  title: Zendvo Auth API
  version: 1.0.0
  description: API for user registration and authentication

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user record with "unverified" status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "StrongP@ss123"
                name:
                  type: string
                  example: "John Doe"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      userId:
                        type: string
                        format: uuid
                      email:
                        type: string
        "400":
          description: Validation error (invalid email, weak password, missing fields)
        "409":
          description: Email already registered
        "413":
          description: Request body too large
        "429":
          description: Too many registration attempts
        "500":
          description: Internal server error
  /api/auth/forgot-password:
    post:
      summary: Request a password reset
      description: Sends a password reset email if the user exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        "200":
          description: Success response (always returned for security)
        "400":
          description: Invalid email format
        "429":
          description: Too many requests
        "500":
          description: Internal server error

  /api/auth/reset-password:
    post:
      summary: Reset user password
      description: Resets the password using a valid reset token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "NewStrongP@ss123"
      responses:
        "200":
          description: Password reset successfully
        "400":
          description: Invalid token or weak password
        "500":
          description: Internal server error
  /api/auth/login:
    post:
      summary: User login
      description: Returns access and refresh tokens for valid credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: "Password123!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
        "401":
          description: Invalid email or password
        "429":
          description: Too many failed attempts from the same IP
        "500":
          description: Internal server error

  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: Returns a new access token using a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Tokens refreshed
        "401":
          description: Invalid or expired refresh token

  /api/auth/logout:
    post:
      summary: User logout
      description: Invalidates the provided refresh token.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Logged out successfully
        "400":
          description: Missing refresh token
        "401":
          description: Unauthorized

  /api/auth/resend-otp:
    post:
      summary: Resend email verification OTP
      description: Resends a verification code to the authenticated user with a 60 second cooldown.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OTP resent successfully or already verified
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "429":
          description: Rate limit exceeded
        "500":
          description: Internal server error

  /api/auth/me:
    get:
      summary: Get current user
      description: Returns the authenticated user's profile and verification status.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                        format: email
                      name:
                        type: string
                        nullable: true
                      role:
                        type: string
                      status:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      lastLogin:
                        type: string
                        format: date-time
                        nullable: true
                      email_verified:
                        type: boolean
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "500":
          description: Internal server error

  /api/gifts:
    get:
      summary: Get user gifts
      description: Returns a list of gifts for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of gifts
        "401":
          description: Unauthorized
    post:
      summary: Create a new gift
      description: Initiates a gift sending process.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipient
                - amount
                - currency
              properties:
                recipient:
                  type: string
                  format: uuid
                amount:
                  type: number
                currency:
                  type: string
                message:
                  type: string
                template:
                  type: string
      responses:
        "201":
          description: Gift created, pending OTP
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "422":
          description: Validation error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT